Bump version to 2.3.0.1 (re #13)
Note: this version bump does not yet represent a package release, but has
been only done so we can ship GHC 7.8 RC1 with a `win32` package version
distinguishable from what shipped with GHC 7.6.3.

Also I bumped the min-version build-dep on `base` as the recent changes
to the `win32` code-base rely on features only available in `base-4.7.0.0`.