Fix problem with RemoveDirectory
RemoveDirectory API exhibits a strange behaviour when removing a
directory that is open. The directory is deleted and the API returns
success, however calling RemoveDirectory for the parent directory
immediately afterwards fails with "The directory is not empty." error.
It appears that the directory deletion is completed asynchronously. This
behaviour is not documented AFAIK but is easy to reproduce:

        1. Create an empty directory tree c:\1\2
        2. Open c:\1\2 in File Explorer
        3. Run the following program:

                RemoveDirectory("c:\\1\\2");
                RemoveDirectory("c:\\1");

The second RemoveDirectory fails with "The directory is not empty." even
though the directory c:\1\2 has been removed. Adding a short delay
between RemoveDirectory calls makes the second call succeed.