Docs: Clarify that isMinTTY only checks for old MinTTYs
Despite the name, `isMinTTY` is only capable of detecting the presence of old
versions of MinTTY. The name of the function is perhaps misleading in that
regard, but it is widely used enough at this point to where renaming it would
be disruptive. Instead, let's just clarify its intent in the documentation.

Resolves #216.